Create Toast component and convert App to a class
Created a toast message component wrapping the Bootstrap toast
implementation. Also changed the App component to a class because I
wanted to call the toast message from anywhere, so that function and the
associated state needed to live at that top level and this caused the
component to start getting complicated and I felt that the complexity
would be easier to manage in a class.

The Admin panel will need to be converted as well at some point.