Bug fix: Not redirecting back to login from admin
If logging out from the admin route the page didn't automtically
redirect back to the login screen as expected. The problem was that the
order of events was causing the admin route that the user was on to not
exist when the redirect was to happen. The fix was to just make sure the
admin route was always there and trust the PrivateRoute component to do
the redirect properly. We may still need to check for a null state in
the Admin panel and redirect back to home or force a logout.